# apps/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "0x1audio",
    "summary": "Private music catalog with search, chat, and playlists.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chats": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "List Chats",
        "operationId": "listChats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatSummaryOut"
                  },
                  "type": "array",
                  "title": "Response Listchats"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create Chat",
        "description": "Create a chat and persist the first user message (reply via stream).",
        "operationId": "createChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Get Chat",
        "operationId": "getChat",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}/messages": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Append Chat Message",
        "operationId": "appendChatMessage",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendMessageBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendMessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}/reply/stream": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Stream Chat Reply",
        "operationId": "streamChatReply",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}/messages/stream": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Stream Chat Message",
        "operationId": "streamChatMessage",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendMessageBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppendMessageBody": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AppendMessageBody"
      },
      "AppendMessageOut": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageOut"
            },
            "type": "array",
            "title": "Messages"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "type": "array",
            "title": "Tracks"
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/PlaylistChatOut"
            },
            "type": "array",
            "title": "Playlists",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "messages",
          "tracks"
        ],
        "title": "AppendMessageOut"
      },
...
      "ArtistRefOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ArtistRefOut"
      },
...
      "ChatDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageOut"
            },
            "type": "array",
            "title": "Messages"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "type": "array",
            "title": "Tracks",
            "default": []
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/PlaylistChatOut"
            },
            "type": "array",
            "title": "Playlists",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "ChatDetailOut"
      },
      "ChatMessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tool_traces": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ToolTraceOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Traces"
          },
          "track_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Ids"
          },
          "playlist_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playlist Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "title": "ChatMessageOut"
      },
      "ChatSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSummaryOut"
      },
...
      "CreateChatBody": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "CreateChatBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaylistChatOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "theme_colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Theme Colors"
          },
          "track_count": {
            "type": "integer",
            "title": "Track Count"
          },
          "preview_tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackOut"
            },
            "type": "array",
            "title": "Preview Tracks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "track_count"
        ],
        "title": "PlaylistChatOut"
      },
...
      "ToolTraceOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "name",
          "args",
          "ok"
        ],
        "title": "ToolTraceOut"
      },
...
      "TrackOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/TrackStatus"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "indexed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "cover_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Color"
          },
          "has_cover": {
            "type": "boolean",
            "title": "Has Cover",
            "default": false
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "is_instrumental": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Instrumental"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistRefOut"
            },
            "type": "array",
            "title": "Artists",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "original_filename",
          "content_type",
          "size_bytes",
          "status",
          "error_message",
          "imported_at",
          "indexed_at"
        ],
        "title": "TrackOut"
      },
      "TrackStatus": {
        "type": "string",
        "enum": [
          "uploading",
          "queued",
          "indexing",
          "ready",
          "failed"
        ],
        "title": "TrackStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}